pyproject.toml
[project]
name = "k8-expert"
version = "0.0.1"
dependencies = [
    "python-dotenv", 
    "google-adk",
    "jinja2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
